To a little code has flash code ,DATA-2020.08.07

+++ code/myflash.c
@@ -6,17 +6,16 @@ void Flash_ToWriteData(uint16_t addr,uint16_t dat)
 
    FLASH_UnLock();
    FLASH_Erase(FLASH_DATA,addr);
-   FLASH_Write(FLASH_DATA,addr, dat);			
-	//temp = FLASH_Read(FLASH_DATA,addr);		
-	
-	//FLASH_Lock();
+   FLASH_Write(FLASH_DATA,addr, dat);	
+  // FLASH_Lock();		
 }
 
 
 
-uint8_t Flash_ToReadData(uint8_t addr)
+uint8_t Flash_ToReadData(uint16_t addr)
 {
 	volatile uint8_t temp=0;
+//	FLASH_UnLock();
 	temp = FLASH_Read(FLASH_DATA,addr);	
     return temp;
 }
